move window-specific logic into src/windows/verify.rs
